test(NODE-4873): move EJSON bigint tests
